cleaned up custom-form-validation code
